Added a button example
To press a midi note, press the button.
To release a midi note, release the button.
